backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23

# File Upload
python-multipart==0.0.6

# AI and OpenAI Integration
openai==2.3.0

# Excel Processing
openpyxl==3.1.2
pandas==2.2.2

# Data Validation and Settings
pydantic==2.12.2
pydantic-settings==2.11.0

# HTTP Client
httpx==0.28.1

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Guardrails AI (Runtime Safety)
guardrails-ai==0.5.0
